deployment fixes

+++ client/src/components/TimeSlot/TimeSlot.tsx
@@ -11,10 +11,10 @@ export function TimeSlot({
   deleteHandler,
   onClick,
 }: {
-  _id: string
+  _id?: string
   from: string
   to: string
-  variant: string
+  variant?: string
   isDisabled?: boolean
   onClick?: (soltId: string) => Promise<void>
   deleteButton?: boolean
@@ -32,7 +32,7 @@ export function TimeSlot({
           borderColor:
             variant == 'outline' && isDisabled ? 'black.600' : 'brand.600',
         }}
-        onClick={onClick && (() => onClick(_id))}
+        onClick={onClick && (() => _id && onClick(_id))}
         variant={variant}
         rounded="md"
         size="lg"
@@ -49,7 +49,7 @@ export function TimeSlot({
           aria-label="delete-slot"
           ml="0.4rem"
           onClick={() => {
-            if (deleteHandler) {
+            if (deleteHandler && _id) {
               deleteHandler(_id)
             }
           }}
